+++ packages/algob/src/lib/tx.ts
@@ -1,11 +1,23 @@
 import { types as rtypes } from "@algo-builder/runtime";
 import { tx as webTx, types as wtypes } from "@algo-builder/web";
-import algosdk, { Algodv2, SuggestedParams, Transaction } from "algosdk";
+import algosdk, { Algodv2, decodeSignedTransaction, SuggestedParams, Transaction } from "algosdk";
 
 import { ConfirmedTxInfo, Deployer } from "../types";
 import { ALGORAND_MIN_TX_FEE } from "./algo-operator";
-import { loadSignedTxnFromFile } from "./files";
+import { loadEncodedTxFromFile } from "./files";
 import { registerCheckpoints } from "./script-checkpoints";
+
+/**
+ * Returns true if encoded transaction (fetched from file) is already signed
+ * @param encodedTx msgpack encoded transaction */
+export function isSignedTx (encodedTx: Uint8Array): boolean {
+  try {
+    decodeSignedTransaction(encodedTx);
+  } catch (error) {
+    return false;
+  }
+  return true;
+}
 
 /**
  * Returns blockchain transaction suggested parameters (firstRound, lastRound, fee..)
@@ -217,6 +229,42 @@ async function mkTx (
 }
 
 /**
+ * Create and Sign SDK transaction(s) from transaction execution parameters (passed by user).
+ * @param deployer Deployer object
+ * @param execParams Execution parameters
+ * @param txIdxMap Map for index to [cpname, asaDef]
+ * @returns [transaction(s), signed transaction(s)]
+ */
+export async function makeAndSignTx (
+  deployer: Deployer,
+  execParams: wtypes.ExecParams | wtypes.ExecParams[],
+  txIdxMap: Map<number, [string, wtypes.ASADef]>
+): Promise<[Transaction[], Uint8Array | Uint8Array[]]> {
+  let signedTxn;
+  let txns: Transaction[] = [];
+  if (Array.isArray(execParams)) {
+    if (execParams.length > 16) { throw new Error("Maximum size of an atomic transfer group is 16"); }
+
+    for (const [idx, txn] of execParams.entries()) {
+      txns.push(await mkTx(deployer, txn, idx, txIdxMap));
+    }
+
+    txns = algosdk.assignGroupID(txns);
+    signedTxn = txns.map((txn: Transaction, index: number) => {
+      const signed = signTransaction(txn, execParams[index]);
+      deployer.log(`Signed transaction ${index}`, signed);
+      return signed;
+    });
+  } else {
+    const txn = await mkTx(deployer, execParams, 0, txIdxMap);
+    signedTxn = signTransaction(txn, execParams);
+    deployer.log(`Signed transaction:`, signedTxn);
+    txns = [txn];
+  }
+  return [txns, signedTxn];
+}
+
+/**
  * Execute single transaction or group of transactions (atomic transaction)
  * @param deployer Deployer
  * @param execParams transaction parameters or atomic transaction parameters
@@ -228,28 +276,8 @@ export async function executeTransaction (
   Promise<ConfirmedTxInfo> {
   deployer.assertCPNotDeleted(execParams);
   try {
-    let signedTxn;
-    let txns: Transaction[] = [];
     const txIdxMap = new Map<number, [string, wtypes.ASADef]>();
-    if (Array.isArray(execParams)) {
-      if (execParams.length > 16) { throw new Error("Maximum size of an atomic transfer group is 16"); }
-
-      for (const [idx, txn] of execParams.entries()) {
-        txns.push(await mkTx(deployer, txn, idx, txIdxMap));
-      }
-
-      txns = algosdk.assignGroupID(txns);
-      signedTxn = txns.map((txn: Transaction, index: number) => {
-        const signed = signTransaction(txn, execParams[index]);
-        deployer.log(`Signed transaction ${index}`, signed);
-        return signed;
-      });
-    } else {
-      const txn = await mkTx(deployer, execParams, 0, txIdxMap);
-      signedTxn = signTransaction(txn, execParams);
-      deployer.log(`Signed transaction:`, signedTxn);
-      txns = [txn];
-    }
+    const [txns, signedTxn] = await makeAndSignTx(deployer, execParams, txIdxMap);
     const confirmedTx = await deployer.sendAndWait(signedTxn);
     console.log(confirmedTx);
     if (deployer.isDeployMode) { await registerCheckpoints(deployer, txns, txIdxMap); }
@@ -272,7 +300,7 @@ export async function executeTransaction (
 export async function executeSignedTxnFromFile (
   deployer: Deployer,
   fileName: string): Promise<ConfirmedTxInfo> {
-  const signedTxn = loadSignedTxnFromFile(fileName);
+  const signedTxn = loadEncodedTxFromFile(fileName);
   if (signedTxn === undefined) { throw new Error(`File ${fileName} does not exist`); }
 
   console.debug("Decoded txn from %s: %O", fileName, algosdk.decodeSignedTransaction(signedTxn));